# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Gateway Plus API
  version: 0.1.0
  description: |
    Personal Gateway Plus 一周可用初版接口契约（P0）。

    设计要点：
    - 对外中转入口 `POST /v1/chat/completions` 保持 OpenAI Chat Completions 兼容，
      `model` 字段填分组 routeKey；仅支持非流式（stream=false）。
    - 成功响应通过响应头 `X-Request-Id`、`X-Upstream` 返回请求 ID 和最终上游名称。
    - 管理接口统一前缀 `/api/admin`；Android 兼容查询接口无前缀（字段名与现有客户端一致）。
    - API Key 永不出现在响应、日志或错误中，仅返回后四位 `apiKeyLastFour`。
    - 所有错误使用统一格式 `ErrorEnvelope`。
  contact:
    name: TSAI Team
servers:
  - url: http://127.0.0.1:8000
    description: 本地开发

# ============================================================
# 全局安全：管理接口暂用简单令牌（初版单用户），中转入口无需鉴权。
# ============================================================
...
paths:
...
  /api/admin/models:
    get:
      tags: [catalog]
      summary: 模型目录列表
      parameters:
        - name: providerId
          in: query
          schema:
            type: string
        - name: includeDisabled
          in: query
          schema:
            type: boolean
            default: false
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelCatalogEntry'
    post:
      tags: [catalog]
      summary: 新增模型目录项
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelCatalogCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelCatalogEntry'
  /api/admin/models/{modelId}:
    parameters:
      - name: modelId
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [catalog]
      summary: 更新模型目录项
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelCatalogUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelCatalogEntry'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [catalog]
      summary: 删除模型目录项及其价格快照
      responses:
        '204':
          description: 删除成功
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/prices:
    get:
      tags: [catalog]
      summary: 价格快照列表（含来源与更新时间）
      parameters:
        - name: providerId
          in: query
          schema:
            type: string
        - name: modelCatalogEntryId
          in: query
          schema:
            type: string
        - name: currentOnly
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceSnapshot'
    post:
      tags: [catalog]
      summary: 发布新的当前价格快照
      description: 同一模型原有的当前快照会自动归档。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceSnapshotCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceSnapshot'
  /api/admin/prices/history/{modelId}:
    get:
      tags: [catalog]
      summary: 查询指定模型的全部价格历史
      parameters:
        - name: modelId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceSnapshot'
  /api/admin/prices/{priceId}:
    delete:
      tags: [catalog]
      summary: 删除价格快照
      description: 删除当前快照时，系统会恢复该模型最近的一条历史快照为当前价格。
      parameters:
        - name: priceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 删除成功
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/promotions:
    get:
      tags: [catalog]
      summary: 活动列表（区分有效/过期）
      parameters:
        - name: providerId
          in: query
          schema:
            type: string
        - name: activeOnly
          in: query
          schema:
            type: boolean
            default: false
        - name: lifecycleStatus
          in: query
          schema:
            type: string
            enum: [draft, upcoming, active, expired]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Promotion'
    post:
      tags: [catalog]
      summary: 新增活动
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
  /api/admin/promotions/{promotionId}:
    parameters:
      - name: promotionId
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [catalog]
      summary: 更新活动
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Promotion'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [catalog]
      summary: 删除活动
      responses:
        '204':
          description: 删除成功
        '404':
          $ref: '#/components/responses/NotFound'

  # ------------------------------------------------------------
  # 路由记录
  # ------------------------------------------------------------
...
  /models:
    get:
      tags: [android-compat]
      summary: Android 兼容 - 模型价格（字段与 LlmModel 一致）
      parameters:
        - name: providerId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AndroidLlmModel'
...
components:
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, type, message]
          properties:
            code:
              type: integer
              description: HTTP 状态码
            type:
              type: string
              description: 错误类别枚举
              enum:
                - validation_error
                - not_found
                - conflict
                - upstream_error
                - all_upstreams_failed
                - request_timeout
                - stream_not_supported
                - internal_error
            message:
              type: string
            requestId:
              type: string
              description: 关联请求 ID
            details:
              type: array
              items:
                type: object

    # ---------- 供应商 ----------
...
    ModelCatalogEntry:
      type: object
      properties:
        id:
          type: string
        providerId:
          type: string
        upstreamModelId:
          type: string
        displayName:
          type: string
        contextWindow:
          type: integer
        enabled:
          type: boolean
        sourceUrl:
          type: string
        verifiedAt:
          type: string
          format: date
    ModelCatalogCreate:
      type: object
      required: [providerId, upstreamModelId, displayName]
      properties:
        providerId:
          type: string
        upstreamModelId:
          type: string
        displayName:
          type: string
        contextWindow:
          type: integer
          minimum: 1
        enabled:
          type: boolean
          default: true
        sourceUrl:
          type: string
          format: uri
        verifiedAt:
          type: string
          format: date-time
    ModelCatalogUpdate:
      type: object
      properties:
        upstreamModelId:
          type: string
        displayName:
          type: string
        contextWindow:
          type: integer
          minimum: 1
        enabled:
          type: boolean
        sourceUrl:
          type: string
          format: uri
          nullable: true
        verifiedAt:
          type: string
          format: date-time
          nullable: true
    PriceSnapshot:
      type: object
      properties:
        id:
          type: string
        providerId:
          type: string
        modelCatalogEntryId:
          type: string
        currency:
          type: string
          default: CNY
        inputPricePerMillionTokens:
          type: number
        outputPricePerMillionTokens:
          type: number
        sourceUrl:
          type: string
        effectiveFrom:
          type: string
          format: date
        verifiedAt:
          type: string
          format: date-time
        isCurrent:
          type: boolean
          description: 是否为该模型当前生效的价格
    PriceSnapshotCreate:
      type: object
      required: [modelCatalogEntryId, currency, sourceUrl, effectiveFrom, verifiedAt]
      properties:
        modelCatalogEntryId:
          type: string
        currency:
          type: string
          minLength: 3
          maxLength: 3
        inputPricePerMillionTokens:
          type: number
          minimum: 0
          nullable: true
        outputPricePerMillionTokens:
          type: number
          minimum: 0
          nullable: true
        sourceUrl:
          type: string
          format: uri
        effectiveFrom:
          type: string
          format: date-time
        verifiedAt:
          type: string
          format: date-time
    Promotion:
      type: object
      properties:
        id:
          type: string
        providerId:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [discount, credit, price_change]
        description:
          type: string
        sourceUrl:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        active:
          type: boolean
          description: 当前是否在有效期内
        status:
          type: string
          enum: [draft, verified, expired]
        lifecycleStatus:
          type: string
          enum: [draft, upcoming, active, expired]
        verifiedAt:
          type: string
          format: date-time
          nullable: true
    PromotionCreate:
      type: object
      required: [providerId, title, type]
      properties:
        providerId:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [discount, credit, price_change]
        description:
          type: string
          nullable: true
        sourceUrl:
          type: string
          format: uri
          nullable: true
        startsAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [draft, verified, expired]
          default: draft
        verifiedAt:
          type: string
          format: date-time
          nullable: true
    PromotionUpdate:
      type: object
      properties:
        providerId:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [discount, credit, price_change]
        description:
          type: string
          nullable: true
        sourceUrl:
          type: string
          format: uri
          nullable: true
        startsAt:
          type: string
          format: date-time
          nullable: true
        endsAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [draft, verified, expired]
        verifiedAt:
          type: string
          format: date-time
          nullable: true

    # ---------- API 分组 ----------
...
    AndroidLlmModel:
      type: object
      description: 对齐 LlmModel.kt
      properties:
        id:
          type: string
        providerId:
          type: string
        name:
          type: string
        contextWindow:
          type: integer
          nullable: true
        inputPricePerMillionTokens:
          type: number
          nullable: true
        outputPricePerMillionTokens:
          type: number
          nullable: true
        currency:
          type: string
          default: USD
        tier:
          type: string
          default: standard
        priceSourceUrl:
          type: string
          nullable: true
        updatedAt:
          type: string
          nullable: true
        priceNote:
          type: string
          nullable: true
...